Extract triplox-client as a publishable workspace member
Splits the HTTP/2 client out of the root triplox crate so it can be
published to crates.io independently of the server. Renames the `edn`
package to `triplox-edn` and uses Cargo's package-rename feature so the
28 `use edn::...` sites in the server compile unchanged.

The client crate owns the wire format (msgpack_codec, protocol) and the
shared client types (DataType, EntityRef, TxOp, QueryArg, TxKey,
TransactionResult) plus the trait surface (SubmitNode, QueryNode,
Database, IntoQuery, IntoTxOp). The server depends on it via path and
re-exports the modules, so existing public paths like
`triplox::client::ClientNode` and `triplox::node::Database` keep
resolving.

Both new crates are versioned 0.1.0-alpha.1.
