+++ Cargo.toml
@@ -1,11 +1,17 @@
 [workspace]
-members = [".", "edn"]
+members = [".", "edn", "triplox-client"]
 exclude = [".claude"]
 
 [workspace.dependencies]
+anyhow = "1.0"
+bytes = { version = "1", features = ["serde"] }
 chrono = "0.4.38"
+reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "http2"] }
+rmp = "0.8"
+rmpv = "1.3"
 serde = "1.0"
 serde_derive = "1.0"
+thiserror = "2.0.1"
 uuid = "1.11.0"
 
 [package]
@@ -14,11 +20,12 @@ version = "0.1.0-beta"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0"
+anyhow = { workspace = true }
 bincode = "1.3"
-bytes = { version = "1", features = ["serde"] }
+bytes = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
-edn = { path = "./edn", features = ["serde_support"] }
+edn = { package = "triplox-edn", path = "./edn", version = "0.1.0-alpha.1", features = ["serde_support"] }
+triplox-client = { path = "./triplox-client", version = "0.1.0-alpha.1" }
 futures = "0.3"
 indexmap = "2.14.0"
 log = "0.4"
@@ -29,7 +36,7 @@ slatedb = { version = "0.12.0", features = ["wal_disable"] }
 # slatedb = { git = "https://github.com/FiV0/slatedb", rev = "5b2f822258e92f682949b1cf7c7aa1c255d4568b", features = ["wal_disable"] }
 slatedb_estimates = { git = "https://github.com/FiV0/slatedb-estimates", rev = "e3ff9dca2a9adc0bf5629c5ff9b2272c7052826c" }
 # slatedb_estimates = { path = "../slatedb-estimates" }
-thiserror = "2.0.1"
+thiserror = { workspace = true }
 time = { version = "0.3.36", features = ["serde"] }
 toml = "0.8"
 tokio = { version = "1.40.0", features = ["full"] }
@@ -42,9 +49,9 @@ axum = "0.8"
 tower = "0.5"
 hyper = { version = "1", features = ["http2", "server"] }
 hyper-util = { version = "0.1", features = ["tokio", "http2"] }
-reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "http2"] }
-rmp = "0.8"
-rmpv = "1.3"
+reqwest = { workspace = true }
+rmp = { workspace = true }
+rmpv = { workspace = true }
 
 [features]
 test-helpers = []
