+++ triplox-client/README.md
@@ -1,0 +1,3 @@
+# triplox-client
+
+HTTP/2 client for [Triplox](https://github.com/FiV0/triplox), a Datalog database.
