+++ triplox-client/src/query.rs
@@ -1,0 +1,3 @@
+use crate::ops::DataType;
+
+pub type QueryResult = Vec<Vec<DataType>>;
